requirements.txt
arrow==1.3.0
asgiref==3.8.1
blessed==1.20.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==41.0.3
Django==5.0.6
django-channels==0.7.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-picklefield==3.2
django-q2==1.7.2
djangogrpcframework==0.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
google-auth==2.40.3
grpcio==1.64.1
grpcio-tools==1.64.1
idna==3.7
inflection==0.5.1
multidict==6.0.5
oauthlib==3.2.2
packaging==24.1
pillow>=10.3.0
protobuf==5.27.1
psycopg2-binary
pulsar-client
pycparser==2.22
PyJWT==2.9.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==3.5.3
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.0
types-python-dateutil==2.9.0.20240821
uritemplate==4.1.1
urllib3==2.2.2
uWSGI==2.0.26
wcwidth==0.2.13
